% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LimeRick.R
\docType{package}
\name{LimeRick}
\alias{LimeRick}
\alias{LimeRick-package}
\alias{lsSessionCache}
\title{Bridge between LimeSurvey and R}
\format{An object of class \code{environment} of length 1.}
\usage{
lsSessionCache
}
\description{
The \code{LimeRick} package ...todo...
}
\details{
If you need help with your research or commercial projects,
feel free to contat me via my homepage contact form:
\url{http://www.wais.kamil.rzeszow.pl}.
}
\seealso{
\itemize{
  \item \url{https://github.com/kalimu/LimeRick} [source code of the latest development version of the R package]
  \item \url{http://limesurvey.org} [LimeSurvey homepage]
}
}
\keyword{datasets}

