% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsGetAnswerOptions.R
\name{ls_getAnswerOptions}
\alias{ls_getAnswerOptions}
\title{Get all answer options for a question in a survey}
\usage{
ls_getAnswerOptions(
  questionID,
  lang = NULL,
  lsAPIurl = getOption("lsAPIurl"),
  sessionKey = NULL
)
}
\arguments{
\item{questionID}{ID of the question}

\item{lang}{\emph{(optional)} Language code for the language - if not given the base language of the particular survey is used}

\item{lsAPIurl}{\emph{(optional)} The URL of the \emph{LimeSurvey RemoteControl 2} JSON-RPC API}

\item{sessionKey}{\emph{(optional)} Authentication token, see \code{\link{lsSessionKey()}}}
}
\value{
A table of answer options
}
\description{
(This convenience function is not directly mapped to a remote procedure.)
}
\examples{
\dontrun{
  lsGetQuestionProperties("13", lang = "en")
  lsGetQuestionProperties("13", lang = "en", properties = list("mandatory"))
}

}
\seealso{
\link{\code{lsGetQuestionProperties()}}
}
