% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsAddFeedback.R
\name{lsAddFeedback}
\alias{lsAddFeedback}
\title{Function for adding new feedback for the package
todo: documentation}
\usage{
lsAddFeedback(feedback, sector, country, packageName = "LimeRick",
  usageStats = getOption("LimeRickStats"))
}
\description{
Function for adding new feedback for the package
todo: documentation
}

