% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsAddResponse.R
\name{lsAddResponse}
\alias{lsAddResponse}
\title{Function for adding new responses to the survey
todo: documentation}
\usage{
lsAddResponse(lsAPIurl, sessionKey, surveyID, response, usageStats = TRUE)
}
\description{
Function for adding new responses to the survey
todo: documentation
}

