% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsAPI.R
\name{lsAPI}
\alias{lsAPI}
\title{Function direclty calling the LimeSurvey API
todo: documentation}
\usage{
lsAPI(lsAPIurl, method, params = NULL)
}
\description{
Function direclty calling the LimeSurvey API
todo: documentation
}

