% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsExportResponses.R
\name{lsExportResponses}
\alias{lsExportResponses}
\title{Export responses}
\usage{
lsExportResponses(
  surveyID,
  lang = NULL,
  completionStatus = "all",
  headingType = "code",
  responseType = "short",
  lsAPIurl = getOption("lsAPIurl"),
  sessionKey = NULL
)
}
\arguments{
\item{surveyID}{ID of the survey}

\item{lang}{\emph{(optional)} Language code for the language to be used (This affects headers if \code{headingType} is not \code{"code"})}

\item{completionStatus}{\emph{(optional)} Response completion status. One out of \code{"complete"}, \code{"incomplete"}, \code{"all"}}

\item{headingType}{\emph{(optional)} Type of the column headers (mainly questions). One out of \code{"code"}, \code{"full"}, \code{"abbreviated"}}

\item{responseType}{\emph{(optional)} Answer codes (\code{"short"}) where applicable, or full answer texts (\code{"long"})}

\item{lsAPIurl}{\emph{(optional)} The URL of the \emph{LimeSurvey RemoteControl 2} JSON-RPC API}

\item{sessionKey}{\emph{(optional)} Authentication token, see \code{\link{lsSessionKey()}}}
}
\value{
A data frame with the survey responses
}
\description{
Export responses
}
\examples{
\dontrun{
  lsExportResponses("123456", completionStatus = "incomplete", headingType = "abbreviated")
}

}
\seealso{
\url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_export_responses}
}
