% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsGetAnswerOptions.R
\name{lsGetAnswerOptions}
\alias{lsGetAnswerOptions}
\title{Function for getting answer options for the question
todo: documentation}
\usage{
lsGetAnswerOptions(surveyID, questionID, lsAPIurl = getOption("lsAPIurl"),
  sessionKey = NULL, usageStats = getOption("LimeRickStats"))
}
\description{
Function for getting answer options for the question
todo: documentation
}

