% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsGetLanguageProperties.R
\name{lsGetLanguageProperties}
\alias{lsGetLanguageProperties}
\title{Get survey language properties}
\usage{
lsGetLanguageProperties(
  surveyID,
  lang = NULL,
  properties = NULL,
  lsAPIurl = getOption("lsAPIurl"),
  sessionKey = NULL
)
}
\arguments{
\item{surveyID}{ID of the Survey}

\item{lang}{\emph{(optional)} Language code for the language to update - if not given the base language of the particular survey is used}

\item{properties}{array|null  \emph{(optional)} A list with the particular properties to get, default to all attributes}

\item{lsAPIurl}{\emph{(optional)} The URL of the \emph{LimeSurvey RemoteControl 2} JSON-RPC API}

\item{sessionKey}{\emph{(optional)} Authentication token, see \code{\link{lsGetSessionKey()}}}
}
\value{
A list in case of success containing the requested values
}
\description{
Get survey language properties
}
\examples{
\dontrun{
  lsGetLanguageProperties("123456")
  lsGetLanguageProperties("123456", lang = "fr")
  lsGetLanguageProperties("123456", properties = list("surveyls_email_confirm"))
  lsGetLanguageProperties("123456", lang = "fr",
                          properties = list("surveyls_email_confirm"))
}

}
\seealso{
\itemize{
  \item \code{\link{lsSetLanguageProperties()}}
  \item \url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_get_language_properties}
}
}
