% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsGetPackageStats.R
\name{lsGetPackageStats}
\alias{lsGetPackageStats}
\title{Function for getting stats of using package functions
todo: documentation}
\usage{
lsGetPackageStats(packageName = "LimeRick",
  usageStats = getOption("LimeRickStats"))
}
\description{
Function for getting stats of using package functions
todo: documentation
}

