% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsGetProperties.R
\name{lsGetProperties}
\alias{lsGetProperties}
\title{Function for getting question or survey properties
todo: documentation}
\usage{
lsGetProperties(lsAPIurl, sessionKey, action = "survey", surveyID = NULL,
  questionID = NULL, lang = "en")
}
\description{
Function for getting question or survey properties
todo: documentation
}

