% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsGetQuestionProperties.R
\name{lsGetQuestionProperties}
\alias{lsGetQuestionProperties}
\title{Get properties of a question in a survey}
\usage{
lsGetQuestionProperties(
  questionID,
  languageCode = NULL,
  properties = NULL,
  lsAPIurl = getOption("lsAPIurl"),
  sessionKey = NULL
)
}
\arguments{
\item{properties}{\emph{(optional)} A vector with the particular property names to request, otherwise get all settings}

\item{lsAPIurl}{\emph{(optional)} The URL of the \emph{LimeSurvey RemoteControl 2} JSON-RPC API}

\item{sessionKey}{\emph{(optional)} Authentication token, see \code{\link{lsSessionKey()}}}

\item{surveyID}{ID of the survey}

\item{lang}{\emph{(optional)} Language code for the language - if not given the base language of the particular survey is used}
}
\value{
A list of question properties or a single property
}
\description{
Get properties of a survey. All internal properties of a survey are available.
}
\examples{
\dontrun{
  lsGetQuestionProperties("13", lang = "en")
  lsGetQuestionProperties("13", lang = "en", properties = list("mandatory"))
}

}
\seealso{
\itemize{
  \item \url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_get_question_properties}
  \item \url{https://api.limesurvey.org/classes/Question.html} (for a list of available properties)
}
}
