% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsGetResponses.R
\name{lsGetResponses}
\alias{lsGetResponses}
\title{Function for getting responses from a survey via LimeSurvey API
todo: documentation}
\usage{
lsGetResponses(surveyID, documentType = "csv", languageCode = "en",
  completionStatus = "all", headingType = "code", responseType = "long",
  fromResponseID = NULL, toResponseID = NULL, fields = NULL,
  lsAPIurl = getOption("lsAPIurl"), sessionKey = NULL,
  usageStats = getOption("LimeRickStats"))
}
\description{
Function for getting responses from a survey via LimeSurvey API
todo: documentation
}

