% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsGetSessionKey.R
\name{lsGetSessionKey}
\alias{lsGetSessionKey}
\title{Create and return a session key}
\usage{
lsGetSessionKey(
  lsAPIurl = getOption("lsAPIurl"),
  username = getOption("lsUser"),
  password = getOption("lsPass"),
  setEnvir = TRUE
)
}
\arguments{
\item{lsAPIurl}{\emph{(optional)} The URL of the \emph{LimeSurvey RemoteControl 2} JSON-RPC API}

\item{username}{\emph{(optional)} The LimeSurvey user name}

\item{password}{\emph{(optional)} The LimeSurvey password}

\item{setEnvir}{Sets a session key environment variable so that it doesn't have to be passed as an argument to other functions}
}
\value{
\itemize{
  \item On success: the session key.
  \item On error: \itemize{
    \item For protocol-level errors (invalid format etc), an error message.
    \item For invalid user name and password, returns a null error and the result body contains a 'status' name-value pair with the error message.
}}
}
\description{
Creates a new JSON-RPC session key. This is mandatory for all following \emph{LimeSurvey RemoteControl 2} function calls.
}
\examples{
\dontrun{
  lsGetSessionKey()
}

}
\references{
\url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_get_session_key}
}
\seealso{
\code{\link{lsReleaseSessionKey}()}
}
