% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsGetSummary.R
\name{lsGetSummary}
\alias{lsGetSummary}
\title{Function for getting survey summary
todo: documentation}
\usage{
lsGetSummary(lsAPIurl, sessionKey, surveyID, usageStats = TRUE)
}
\description{
Function for getting survey summary
todo: documentation
}

