% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsGetSurveyProperties.R
\name{lsGetSurveyProperties}
\alias{lsGetSurveyProperties}
\title{Get survey properties}
\usage{
lsGetSurveyProperties(
  surveyID,
  properties = NULL,
  lsAPIurl = getOption("lsAPIurl"),
  sessionKey = NULL
)
}
\arguments{
\item{surveyID}{ID of the survey}

\item{properties}{\emph{(optional)} A vector with the particular property names to request, otherwise get all settings}

\item{lsAPIurl}{\emph{(optional)} The URL of the \emph{LimeSurvey RemoteControl 2} JSON-RPC API}

\item{sessionKey}{\emph{(optional)} Authentication token, see \code{\link{lsSessionKey()}}}
}
\value{
A list of survey properties or a single property
}
\description{
Get properties of a survey. All internal properties of a survey are available.
}
\examples{
\dontrun{
  lsGetSurveyProperties("123456")
  lsGetSurveyProperties("123456", properties = list("anonymized"))
  lsGetSurveyProperties("123456", properties = list("adminemail",
                                                    "anonymized"))
}

}
\seealso{
\itemize{
  \item \url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_get_survey_properties}
  \item \url{https://api.limesurvey.org/classes/Survey.html} (for a list of available properties)
}
}
