% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsListGroups.R
\name{lsListGroups}
\alias{lsListGroups}
\title{Get survey groups}
\usage{
lsListGroups(
  surveyID,
  lang = NULL,
  lsAPIurl = getOption("lsAPIurl"),
  sessionKey = NULL
)
}
\arguments{
\item{surveyID}{ID of the survey containing the groups}

\item{lang}{\emph{(optional)} Language code for the language -- if not given, due to a bug in the API, instead of just the base language of the particular survey, all languages are returned}

\item{lsAPIurl}{\emph{(optional)} The URL of the \emph{LimeSurvey RemoteControl 2} JSON-RPC API}

\item{sessionKey}{\emph{(optional)} Authentication token, see \code{\link{lsGetSessionKey()}}}
}
\value{
The ids and all attributes of all survey groups
}
\description{
Get survey groups
}
\examples{
\dontrun{
  lsListGroups("123456")
  lsListGroups("123456", lang = "fr")
}

}
\references{
\url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_list_groups}
}
