% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsListQuestions.R
\name{lsListQuestions}
\alias{lsListQuestions}
\title{Return the ids and info of (sub-)questions of a survey/group}
\usage{
lsListQuestions(
  surveyID,
  lang,
  groupID = NULL,
  lsAPIurl = getOption("lsAPIurl"),
  sessionKey = NULL
)
}
\arguments{
\item{surveyID}{ID of the survey}

\item{lang}{Language code for the survey language (\strong{Note:} The API expects
one of the survey languages as part of the request rather than falling back to
the default language of the survey. However, you can look up that default
language using \code{\link{lsGetSurveyProperties()}})}

\item{groupID}{\emph{(optional)} ID of the group to list questions}

\item{lsAPIurl}{\emph{(optional)} The URL of the \emph{LimeSurvey RemoteControl 2} JSON-RPC API}

\item{sessionKey}{\emph{(optional)} Authentication token, see \code{\link{lsGetSessionKey()}}}
}
\value{
The list of questions
}
\description{
Return the ids and info of (sub-)questions of a survey/group
}
\examples{
\dontrun{
  lsListQuestions("123456", lang = "fr")
  lsListQuestions("123456", lang = "fr", groupID = 1)
}

}
\seealso{
\url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_list_questions}
}
