% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsListSurveys.R
\name{lsListSurveys}
\alias{lsListSurveys}
\title{List available surveys}
\usage{
lsListSurveys(lsAPIurl = getOption("lsAPIurl"), sessionKey = NULL)
}
\arguments{
\item{lsAPIurl}{\emph{(optional)} The URL of the \emph{LimeSurvey RemoteControl 2} JSON-RPC API}

\item{sessionKey}{\emph{(optional)} Authentication token, see \code{\link{lsGetSessionKey}()}}
}
\value{
An dataframe with
  \code{sid} the ids of survey,
  \code{surveyls_title} the title of the survey,
  \code{startdate} start date,
  \code{expires} expiration date,
  \code{active} if survey is active (Y) or not (!Y)
}
\description{
List available surveys
}
\examples{
\dontrun{
  lsListSurveys()
}

}
\references{
\url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_list_surveys}
}
