% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LimeRick.R
\docType{data}
\name{lsSessionCache}
\alias{lsSessionCache}
\title{Setting a new environment to hold the session parameters.}
\format{
An object of class \code{environment} of length 1.
}
\usage{
lsSessionCache
}
\description{
The \code{lsSessionCache} environment is crated upon the package loading
and stores session data like session key or timestamp when the session started.
The environment is accessed directly only by package functions not by a package user.
}
\references{
\url{https://github.com/cloudyr/limer/blob/master/R/get_session_key.R}
}
\keyword{internal}
