% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsSessionKey.R
\name{lsSessionKey}
\alias{lsSessionKey}
\title{Function getting and releasing LimeSurvey API session key
todo: documentation}
\usage{
lsSessionKey(lsAPIurl, action = "release", user = NULL, pass = NULL)
}
\description{
Function getting and releasing LimeSurvey API session key
todo: documentation
}

