% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsSetLanguageProperties.R
\name{lsSetLanguageProperties}
\alias{lsSetLanguageProperties}
\title{Set survey language properties}
\usage{
lsSetLanguageProperties(
  surveyID,
  properties,
  lang = NULL,
  lsAPIurl = getOption("lsAPIurl"),
  sessionKey = NULL
)
}
\arguments{
\item{surveyID}{ID of the survey}

\item{properties}{A list with the particular field names as keys and their values to set on that particular survey}

\item{lang}{\emph{(optional)} Language code for the language to update - if not given the base language of the particular survey is used}

\item{lsAPIurl}{\emph{(optional)} The URL of the \emph{LimeSurvey RemoteControl 2} JSON-RPC API}

\item{sessionKey}{\emph{(optional)} Authentication token, see \code{\link{lsGetSessionKey()}}}
}
\value{
\code{OK} if save successful otherwise error text
}
\description{
Set survey language properties
}
\examples{
\dontrun{
  lsSetLanguageProperties("123456", properties = list(
    surveyls_email_confirm = "Hallo, this text comes from LimeRick"))

  lsSetLanguageProperties("123456", lang = "fr", properties = list(
    surveyls_email_confirm = "Bonjour, this text comes from LimeRick"))
}

}
\references{
\url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_set_language_properties}
}
\seealso{
\code{\link{lsGetLanguageProperties()}}
}
