% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LimeRick.R
\docType{package}
\name{LimeRick}
\alias{LimeRick}
\title{Bridge between LimeSurvey and R}
\description{
This package is a \emph{\href{https://www.limesurvey.org/}{LimeSurvey} RemoteControl 2} JSON-RPC API Client for R.
}
\details{
This package enables close connection between R and LimeSurvey
— an open source applocation for online interviewing.
Using the LimeSurvey API for bidirectional data exchange allows for:
\itemize{
  \item importing survey responses into R,
  \item adding new responses to the survey directly from R,
  \item automatically accessing surveys and question properties,
  \item developing data products based on real time declarative data collection.
}
}
\references{
\itemize{
  \item \url{https://k127.github.io/LimeRick/} (package homepage)
  \item \url{https://github.com/k127/LimeRick/} (source code)
  \item \url{https://github.com/k127/LimeRick/issues/} (issue tracker)
}
}
