% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LimeRick.R
\docType{package}
\name{LimeRick}
\alias{LimeRick}
\alias{LimeRick-package}
\alias{lsSessionCache}
\title{Bridge between LimeSurvey and R}
\format{An object of class \code{environment} of length 1.}
\usage{
lsSessionCache
}
\description{
The \code{LimeRick} package enables close connection between R and LimeSurvey – one of the most open-source scripts for Computer-Aided Web Interviewing. Using new LimeSurvey RemoteControl2 API for two-way data exchange allows for:
\itemize{
  \item importing on-line survey responses into R,
  \item adding new responses to the survey directly from R,
  \item automatically accessing surveys and question properties,
  \item implementing passive measurement surveys,
  \item developing data products based on real-time declarative data collection.
}
}
\details{
The lsSessionCache environment is crated upon the package loading
and stores session data like session key or timestamp when the session started.
The environment is accessed directly only by package functions not by a package user.

If you need help with your research or commercial projects,
feel free to contat me via my homepage contact form:
\url{http://www.wais.kamil.rzeszow.pl}.
}
\seealso{
\itemize{
  \item \url{http://www.wais.kamil.rzeszow.pl/LimeRick} [the package homepage]
  \item \url{https://github.com/kalimu/LimeRick} [source code & development version of the package]
  \item \url{http://limesurvey.org} [LimeSurvey homepage]
}
}
\keyword{datasets}

