% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsAddPackageStats.R
\name{lsAddPackageStats}
\alias{lsAddPackageStats}
\title{Function for adding new stats of using package functions
todo: documentation}
\usage{
lsAddPackageStats(packageName = "LimeRick", packageVer = NULL, functionName,
  functionStats, usageStats = getOption("LimeRickStats"))
}
\description{
Function for adding new stats of using package functions
todo: documentation
}

