% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsAddResponse.R
\name{lsAddResponse}
\alias{lsAddResponse}
\title{Function for adding new responses to the survey
todo: documentation}
\usage{
lsAddResponse(surveyID, response, lsAPIurl = getOption("lsAPIurl"),
  sessionKey = NULL, usageStats = getOption("LimeRickStats"))
}
\description{
Function for adding new responses to the survey
todo: documentation
}

