% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsAPI.R
\name{lsAPI}
\alias{lsAPI}
\title{Function directly calling the LimeSurvey API
todo: documentation}
\usage{
lsAPI(method, params = NULL, lsAPIurl = getOption("lsAPIurl"))
}
\description{
Function directly calling the LimeSurvey API
todo: documentation
}

