% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsGetProperties.R
\name{lsGetProperties}
\alias{lsGetProperties}
\title{Function for getting question or survey properties
todo: documentation}
\usage{
lsGetProperties(action = "survey", surveyID = NULL, questionID = NULL,
  languageCode = "en", lsAPIurl = getOption("lsAPIurl"),
  sessionKey = NULL, usageStats = getOption("LimeRickStats"))
}
\description{
Function for getting question or survey properties
todo: documentation
}

