% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsGetQuestionProperties.R
\name{lsGetQuestionProperties}
\alias{lsGetQuestionProperties}
\title{Get properties of a question in a survey}
\usage{
lsGetQuestionProperties(
  questionID,
  lang,
  properties = NULL,
  lsAPIurl = getOption("lsAPIurl"),
  sessionKey = NULL
)
}
\arguments{
\item{questionID}{ID of the question}

\item{lang}{Language code for the survey language which affects the question
title amongst other properties (\strong{Note:} The API expects
one of the survey languages as part of the request rather than falling back to
the default language of the survey. However, you can look up that default
language using \code{\link{lsGetSurveyProperties()}})}

\item{properties}{\emph{(optional)} A vector with the particular property names to request, otherwise get all settings}

\item{lsAPIurl}{\emph{(optional)} The URL of the \emph{LimeSurvey RemoteControl 2} JSON-RPC API}

\item{sessionKey}{\emph{(optional)} Authentication token, see \code{\link{lsGetSessionKey()}}}
}
\value{
A list of question properties or a single property
}
\description{
Get properties of a survey. All internal properties of a survey are available.
}
\examples{
\dontrun{
  lsGetQuestionProperties("13", lang = "en")
  lsGetQuestionProperties("13", lang = "en", properties = list("mandatory"))
}

}
\seealso{
\itemize{
  \item \code{\link{lsSetQuestionProperties()}}
  \item \url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_get_question_properties}
  \item \url{https://api.limesurvey.org/classes/Question.html} (for a list of available properties)
}
}
