% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsGetSummary.R
\name{lsGetSummary}
\alias{lsGetSummary}
\title{Function for getting survey summary
todo: documentation}
\usage{
lsGetSummary(surveyID, lsAPIurl = getOption("lsAPIurl"), sessionKey = NULL,
  usageStats = getOption("LimeRickStats"))
}
\description{
Function for getting survey summary
todo: documentation
}

