% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsList.R
\name{lsList}
\alias{lsList}
\title{Function for listing surveys, groups or questions
todo: documentation}
\usage{
lsList(lsAPIurl, sessionKey, action = "surveys", surveyID = NULL,
  usageStats = TRUE)
}
\description{
Function for listing surveys, groups or questions
todo: documentation
}

