% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsListQuestions.R
\name{lsListQuestions}
\alias{lsListQuestions}
\title{Return the ids and info of (sub-)questions of a survey/group}
\usage{
lsListQuestions(
  surveyID,
  groupID = NULL,
  lang = NULL,
  lsAPIurl = getOption("lsAPIurl"),
  sessionKey = NULL
)
}
\arguments{
\item{surveyID}{ID of the survey}

\item{groupID}{\emph{(optional)} ID of the group to list questions}

\item{lang}{\emph{(optional)} Language code for the language - if not given the base language of the particular survey is used}

\item{lsAPIurl}{\emph{(optional)} The URL of the \emph{LimeSurvey RemoteControl 2} JSON-RPC API}

\item{sessionKey}{\emph{(optional)} Authentication token, see \code{\link{lsSessionKey}}}
}
\value{
The list of questions
}
\description{
Return the ids and info of (sub-)questions of a survey/group
}
\examples{
lsListQuestions("123456")
lsListQuestions("123456", lang = "fr")
lsListQuestions("123456", lang = "fr", groupID = 1)

}
\seealso{
\url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_list_questions}
}
