% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsReleaseSessionKey.R
\name{lsReleaseSessionKey}
\alias{lsReleaseSessionKey}
\title{Close the RPC session}
\usage{
lsReleaseSessionKey(lsAPIurl = getOption("lsAPIurl"))
}
\arguments{
\item{lsAPIurl}{\emph{(optional)} The URL of the \emph{LimeSurvey RemoteControl 2} JSON-RPC API}
}
\value{
\code{OK}
}
\description{
Closes a previously opened JSON-RPC session ("Log out").
}
\examples{
\dontrun{
  lsReleaseSessionKey()
}

}
\seealso{
\itemize{
  \item \code{\link{lsGetSessionKey()}}
  \item \url{https://api.limesurvey.org/classes/remotecontrol_handle.html#method_release_session_key}
}
}
