% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsSessionKey.R
\name{lsSessionKey}
\alias{lsSessionKey}
\title{Function getting and releasing LimeSurvey API session key
todo: documentation}
\usage{
lsSessionKey(action = "set", lsAPIurl = getOption("lsAPIurl"),
  user = getOption("lsUser"), pass = getOption("lsPass"), verbose = TRUE,
  sessionType = "global")
}
\description{
Function getting and releasing LimeSurvey API session key
todo: documentation
}

